exhaustive pixel format + audio changes

+++ src/audio/device.cpp
@@ -61,7 +61,7 @@ audio::device::~device()
 
 void audio::device::queue(std::span<std::byte> bytes)
 {
-    HAL_ASSERT_VITAL(::SDL_QueueAudio(m_id, bytes.data(), bytes.size_bytes()) == 0, debug::last_error());
+    HAL_ASSERT_VITAL(::SDL_QueueAudio(m_id, bytes.data(), static_cast<Uint32>(bytes.size_bytes())) == 0, debug::last_error());
 }
 
 void audio::device::pause(bool p)
